Add NSNextOid to keyedobjects.nib to prevent "fixing an internal inconsistency"
The gist is that compiled NIBs don't have an NSNextOid value set in their IB.objectdata dictionary. Xcode, when attempting to open a file without this key, will attempt to repair the file, causing the file to require a save before being closed. When Xcode attempts to save one of these hacky NIB files, however, a crash invariably results. This means that opening any of these files generated by the previous version of Nib-Decompiler would get you into an annoying crash loop, and you'd have to move the NIB to the trash before Xcode would open for more than 10 seconds at a time again.

This is still nowhere near a perfect decompiler, and such a thing is likely not even possible; it's already amazing what this one does with such little code (and one dumb little trick that makes IB think it's opening a legacy NIB document format). However, whenever an NIB uses a newer feature of IB (new as in since ~Xcode 4), they sometimes will simply not open (because the legacy format converter doesn't know how to deal with the new stuff), and at the very least almost always fail to save (due to the existence of autolayout, and once again, legacy converters having no idea what to do with it).